requirements.txt
absl-py==1.4.0
astropy==5.2.2
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
exceptiongroup==1.2.0
Fiona==1.9.4.post1
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
geopandas==0.13.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
gurobipy==11.0.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
isal==1.2.0
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
libpysal==4.8.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib2tikz==0.7.6
matsim-tools==0.0.14
mdurl==0.1.2
momepy==0.6.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==4.0.0
pluggy==1.3.0
POAP==0.1.26
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDOE2==1.3.0
pyerfa==2.0.0.3
Pygments==2.16.1
pyparsing==3.0.9
pyproj==3.5.0
pySOT==0.3.3
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.2.0
tikzplotlib==0.10.1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
webcolors==1.13
Werkzeug==2.3.7
wrapt==1.15.0
xopen==1.7.0
zipp==3.16.2
